Refactor cache module
