+++ groove/proxy/cache/invalidator_test.go
@@ -1,0 +1,38 @@
+package cache
+
+import (
+	"io/ioutil"
+	"testing"
+)
+
+func TestSaveReadIndex(t *testing.T) {
+	cacheDirectory, err := ioutil.TempDir("", "")
+	if err != nil {
+		t.Fatalf("Error creating temp dir: %s", err)
+	}
+
+	invalidator := NewCacheInvalidator(cacheDirectory, 10, 10, 1)
+	invalidator.Set("testKey", &TestSimpleObject{"testValue"})
+
+	// Ensure it saved automatically - should have spanwed a goroutine
+	invalidator.saveWaiter.Wait()
+
+	// Attempt to read the index file
+	fileContents, _, err := invalidator.readIndex()
+
+	if err != nil {
+		t.Fatalf("Error reading index: %s", err)
+	}
+
+	if len(fileContents) != 1 {
+		t.Fatalf("Index should have one entry")
+	}
+
+	if fileContents[0].Key != "testKey" {
+		t.Fatalf("Index should have testKey (actual: %s)", fileContents[0].Key)
+	}
+
+	if fileContents[0].Size == 0 {
+		t.Fatalf("Index should have non-zero size (actual: %d)", fileContents[0].Size)
+	}
+}
