+++ groove/proxy/cache/utilities_test.go
@@ -1,8 +1,12 @@
-package main
+package cache
 
 import (
 	"testing"
 )
+
+type TestSimpleObject struct {
+	Value string
+}
 
 func TestEncodeDecodeObject(t *testing.T) {
 	object := &TestSimpleObject{
