--- groove/proxy/test_shared.go
@@ -1,5 +1,0 @@
-package main
-
-type TestSimpleObject struct {
-	Value string
-}
